// sistrado-app/src/app/model/request/registro-request.ts
export interface RegistroRequest {
    nombres: string;
    apePaterno: string;
    apeMaterno: string;
    celular: string;
    departamento: string;
    provincia: string;
    distrito: string;
    direccion: string;
    dni: string;
    email: string;
    clave: string;
    representante: boolean;
    ruc: string;
    dirFiscal: string;
}